requirements.txt
git+https://github.com/singpenguin/loadcf.git@master
peewee==3.10.0
sanic==19.6.3
python-logstash==0.4.6
